% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggiungi_risposta.R
\name{aggiungi_risposta}
\alias{aggiungi_risposta}
\title{Add answer}
\usage{
aggiungi_risposta(a, domanda, risposta)
}
\arguments{
\item{a}{solution object}

\item{domanda}{\link{int} the number(s) of the question(s) to answer}

\item{risposta}{\link{any} the object(s) with the answer to the question(s)}
}
\value{
a solution object
}
\description{
Add answer
}
\examples{
esame_corrado <- soluzione(
  candidato = list(nome = "Corrado", cognome = "Lanera", mat = 1234),
  corso     = "tdp",
  anno      = 2018,
  appello   = 2,
  tema      = "a"
)
esame_corrado <- aggiungi_risposta(esame_corrado, 2, TRUE)
esame_corrado
}
