% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{death_ita}
\alias{death_ita}
\title{Proportion of deaths, country: Italy}
\format{A data frame with 15 rows and 52 variables:}
\source{
\url{http://apps.who.int/gho/data/view.main-amro.ghe3002015-ITA}
}
\usage{
death_ita
}
\description{
Distribution of causes of death among children aged <5 years (%)
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
