% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibile_risposta.R
\name{possibile_risposta}
\alias{possibile_risposta}
\title{Title}
\usage{
possibile_risposta(answer_obj)
}
\arguments{
\item{answer_obj}{}
}
\description{
Title
}
\examples{
possibile_risposta(answer_1)
answer_1 <- 1
possibile_risposta(answer_1)
}
