% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_tdp1718.R
\name{tdp1718_check_1}
\alias{tdp1718_check_1}
\title{tdp check 1}
\usage{
tdp1718_check_1(file_to_check = "death_ita.csv")
}
\arguments{
\item{file_to_check}{\link{chr} name of a file which should be into the WD}
}
\value{
logical (invisibly)
}
\description{
tdp check 1
}
