% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_tdp1718.R
\name{tdp1718_check_4}
\alias{tdp1718_check_4}
\title{tdp check 4}
\usage{
tdp1718_check_4(main_df = death_ita)
}
\arguments{
\item{main_df}{a data.frame}
}
\value{
logical (invisibly)
}
\description{
tdp check 4
}
