% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_tdp1718.R
\name{tdp1718_check_6}
\alias{tdp1718_check_6}
\title{tdp check 6}
\usage{
tdp1718_check_6(text_var = "causa_del_decesso")
}
\arguments{
\item{text_var}{\link{chr} name of the variable to transform}
}
\value{
logical (invisibly)
}
\description{
tdp check 6
}
