% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_tdp1718.R
\name{tdp1718_check_8}
\alias{tdp1718_check_8}
\title{tdp check 8}
\usage{
tdp1718_check_8(df_name = "death_ita", ord_one = "percentuale",
  ord_two = "anno", ord_three = "mesi")
}
\arguments{
\item{df_name}{\link{chr} the name of the principal data.frame}
}
\value{
logical (invisibly)
}
\description{
tdp check 8
}
