% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_tdp1718.R
\name{tdp1718_check_9}
\alias{tdp1718_check_9}
\title{tdp check 9}
\usage{
tdp1718_check_9(matrix_name = "median_p_death_causes")
}
\arguments{
\item{matrix_name}{\link{chr} the name of the requested matrix}
}
\value{
logical (invisibly)
}
\description{
tdp check 9
}
