% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_2_a_checks.R
\name{tdp_2018_2_a_check_1}
\alias{tdp_2018_2_a_check_1}
\title{tdp 2018, appello 2, tema a, check 1}
\usage{
tdp_2018_2_a_check_1(wd = getwd(), file_to_check = "death_ita.csv")
}
\arguments{
\item{wd}{\link{chr} working directory}

\item{file_to_check}{\link{chr} path of a file to check into the wd}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 2, tema a, check 1
}
