% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_2_a_checks.R
\name{tdp_2018_2_a_check_2}
\alias{tdp_2018_2_a_check_2}
\title{tdp 2018, appello 2, tema a, check 2}
\usage{
tdp_2018_2_a_check_2(provided_names = NULL, required_names = "death_ita")
}
\arguments{
\item{provided_names}{names included in the workspace}

\item{required_names}{variable'names required}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 2, tema a, check 2
}
