% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_2_a_checks.R
\name{tdp_2018_2_a_check_4}
\alias{tdp_2018_2_a_check_4}
\title{tdp 2018, appello 2, tema a, check 4}
\usage{
tdp_2018_2_a_check_4(main_df)
}
\arguments{
\item{main_df}{a data.frame}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 2, tema a, check 4
}
