% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_2_a_checks.R
\name{tdp_2018_2_a_check_6}
\alias{tdp_2018_2_a_check_6}
\title{tdp 2018, appello 2, tema a, check 6}
\usage{
tdp_2018_2_a_check_6(main_df, new_value = "")
}
\arguments{
\item{main_df}{\link{data.frame} main dataframe}

\item{new_value}{\link{chr} character used to substitute the white spaces}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 2, tema a, check 6
}
