% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_2_a_checks.R
\name{tdp_2018_2_a_check_7}
\alias{tdp_2018_2_a_check_7}
\title{tdp 2018, appello 2, tema a, check 7}
\usage{
tdp_2018_2_a_check_7(main_df)
}
\arguments{
\item{main_df}{\link{data.frame} main dataframe}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 2, tema a, check 7
}
