% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_2_a_checks.R
\name{tdp_2018_2_a_check_9}
\alias{tdp_2018_2_a_check_9}
\title{tdp 2018, appello 2, tema a, check 9}
\usage{
tdp_2018_2_a_check_9(main_table)
}
\arguments{
\item{main_table}{\link{matrix} the requested matrix}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 2, tema a, check 9
}
