% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_3_a_checks.R
\name{tdp_2018_3_a_check_3}
\alias{tdp_2018_3_a_check_3}
\title{tdp 2018,appello 3, tema a, check 3}
\usage{
tdp_2018_3_a_check_3(data, reference)
}
\arguments{
\item{data}{a list of data to check}

\item{reference}{a list of real data}
}
\value{
logical (invisibly)
}
\description{
tdp 2018,appello 3, tema a, check 3
}
