% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_3_b_checks.R
\name{tdp_2018_3_b_check_2}
\alias{tdp_2018_3_b_check_2}
\title{tdp 2018, appello 3, tema a, check 2}
\usage{
tdp_2018_3_b_check_2(main_data)
}
\arguments{
\item{main_data}{main data.frame}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 3, tema a, check 2
}
