% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_2018_3_b_checks.R
\name{tdp_2018_3_b_check_5}
\alias{tdp_2018_3_b_check_5}
\title{tdp 2018, appello 3, tema a, check 5}
\usage{
tdp_2018_3_b_check_5(main_df, .fun = toupper)
}
\arguments{
\item{.fun}{the function requested for the transformation. Note, the
function schould be an idempotent one, i.e.
\code{.fun(.fun(x)) == .fun(x)}.`}

\item{text_var}{\link{chr} name of the variable to transform}
}
\value{
logical (invisibly)
}
\description{
tdp 2018, appello 3, tema a, check 5
}
