% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlla_esame.R
\name{controlla_esame}
\alias{controlla_esame}
\alias{controlla_esame.default}
\alias{controlla_esame.tdp_2018_2_a}
\title{controlla esame intero}
\usage{
controlla_esame(soluzione)

\method{controlla_esame}{default}(soluzione)

\method{controlla_esame}{tdp_2018_2_a}(soluzione)
}
\arguments{
\item{soluzione}{an object of class rexams}
}
\value{
the score for the test
}
\description{
controlla esame intero
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default methods

\item \code{tdp_2018_2_a}: default methods
}}

