% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlla_risposta.R
\name{controlla_risposta}
\alias{controlla_risposta}
\alias{controlla_risposta.default}
\alias{controlla_risposta.tdp_2018_2_a}
\title{controlla risposta}
\usage{
controlla_risposta(soluzione, domanda)

\method{controlla_risposta}{default}(soluzione, domanda)

\method{controlla_risposta}{tdp_2018_2_a}(soluzione, domanda)
}
\arguments{
\item{soluzione}{an object of class rexams}

\item{domanda}{the number of question to check}
}
\value{
a the output of the test
}
\description{
controlla risposta
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default methods

\item \code{tdp_2018_2_a}: default methods
}}

